app/requirements.txt
pillow
pybase64
numpy
grpcio==1.64.1
grpcio-tools==1.64.1
huggingface-hub
transformers
# torch